Chronic progressive repeated measures (CPRM) model sample size calculations.

This function performs sample size calculations for the chronic progressive repeated measures (CPRM) model when used to test for differences of change scores between groups at last visit. Input parameters are random effect variance and residual error variance as estimated by a REML fit to representative pilot data or data from a representative prior clinical trial or cohort study.

cprm.power(
  n = NULL,
  delta = NULL,
  power = NULL,
  t = NULL,
  lambda = 1,
  sig2.int = 0,
  sig2.s = NULL,
  sig.b0b1 = 0,
  sig2.e = NULL,
  sig2.int_2 = NULL,
  sig2.s_2 = NULL,
  sig.b0b1_2 = NULL,
  sig2.e_2 = NULL,
  sig.level = 0.05,
  p = NULL,
  p_2 = NULL,
  alternative = c("two.sided", "one.sided"),
  tol = NULL
)

Arguments

n

sample size, group 1

delta

group difference in fixed effect slopes

power

power

t

the observation times

lambda

allocation ratio (sample size group 1 divided by sample size group 2)

sig2.int

variance of random intercepts, group 1

sig2.s

variance of random slopes, group 1

sig.b0b1

covariance of random slopes and intercepts,group 1

sig2.e

residual variance, group 1

sig2.int_2

variance of random intercepts, group 2 (defaults to sig2.int)

sig2.s_2

variance of random slopes, group 2 (defaults to sig2.s)

sig.b0b1_2

covariance of random slopes and intercepts, group 2 (defaults to sig.b0b1)

sig2.e_2

residual variance, group 2 (defaults to sig2.e)

sig.level

type one error

p

proportion vector for group 1, if i indexes visits, p[i] = the proportion whose last visit was at visit i (p sums to 1)

p_2

proportion vector for group 2 (defaults to p)

alternative

one- or two-sided test

tol

not used (no root finding used in this implementation).

Value

One of the number of subject required per arm, the power, or detectable effect size given sig.level and the other parameter estimates.

Details

Default settings perform sample size / power / effect size calculations assuming equal covariance of repeated measures in the 2 groups, equal residual error variance across groups, equal allocation to groups, and assuming no study subject attrition. Specifically, variance parameters required for default settings are sig2.s, the variance of random slopes, and sig2.e, the residual error variance, both either known or estimated from a mixed model fit by REML to prior data.

This function accommodates different variance parameters across groups, unequal allocation across groups, and study subject attrition (loss to followup), which may also vary across groups. Details can be found in the description of edland.linear.power

References

Zhao Y, Edland SD. The chronic progressive repeated measures (CPRM) model for longitudinal data. In process.

See also

Author

Michael C. Donohue, Steven D. Edland, Yu Zhao

Examples

if (FALSE) { # \dontrun{
browseVignettes(package = "longpower")
} # }
# An Alzheimer's Disease example using ADAS-cog pilot estimates
t <- seq(0,1.5,0.25)
cprm.power(delta=1.5, t=t, sig2.s = 24, sig2.e = 10, sig.level=0.05, power = 0.80)
#> 
#>      Zhao and Edland, in process 
#> 
#>               N = 1032.564
#>               n = 516.2819, 516.2819
#>           delta = 1.5
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>             p_2 = 0, 0, 0, 0, 0, 0, 1
#>        sig2.int = 0
#>        sig.b0b1 = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>      sig2.int_2 = 0
#>      sig.b0b1_2 = 0
#>        sig2.s_2 = 24
#>        sig2.e_2 = 10
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group