Random coefficient regression models (RCRM) sample size calculations

This function computes sample size and power needed for the random coefficient regression models (RCRM) based on the formula from Hu, Mackey, and Thomas (2021). The RCRM assumes that the experimental and control arms have the same population baseline value.

hu.mackey.thomas.linear.power(
  n = NULL,
  delta = NULL,
  power = NULL,
  t = NULL,
  lambda = 1,
  sig2.i = 0,
  cor.s.i = NULL,
  sig2.s = 0,
  sig2.e = NULL,
  p = NULL,
  sig.level = 0.05,
  alternative = c("two.sided", "one.sided"),
  tol = .Machine$double.eps^2
)

Arguments

n

sample size, group 1. This formula can accommodate unbalanced group allocation via lambda.

delta

Effect size (absolute difference in rate of decline between tx and placebo)

power

power

t

Vector of visit time points (including time 0)

lambda

allocation ratio (sample size group 1 divided by sample size group 2)

sig2.i

Variance of random intercept

cor.s.i

Correlation between random intercept & slope

sig2.s

Variance of random slope

sig2.e

Variance of pure error

p

proportion vector for both groups; if i indexes visits, p[i] = the proportion whose last visit was at visit i (p sums to 1)

sig.level

type one error

alternative

one- or two-sided test

tol

numerical tolerance used in root finding

Value

One of the number of subject required per arm, the power, or detectable effect size given sig.level and the other parameter estimates.

Details

See Hu. Mackey, and Thomas (2021) for parameter details.

See Equations (7) and (8) in Hu, Mackey, and Thomas (2021)

References

Hu, N., Mackey, H., & Thomas, R. (2021). Power and sample size for random coefficient regression models in randomized experiments with monotone missing data. Biometrical Journal, 63(4), 806-824.

See also

Author

Monarch Shah

Examples

if (FALSE) { # \dontrun{
browseVignettes(package = "longpower")
} # }
# An Alzheimer's Disease example using ADAS-cog pilot estimates
t <- seq(0,1.5,0.25)
p <- c(rep(0, 6),1)

hu.mackey.thomas.linear.power(delta=1.5, t=t, 
  sig2.s=24, sig2.e=10, cor.s.i=0.5, p=p, power=0.80)
#> 
#>      Random Coefficients Regression Model. Hu, Mackey & Thomas (2021) 
#> 
#>               N = 359.4193
#>               n = 179.7096, 179.7096
#>           delta = 1.5
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>          sig2.i = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>         cor.s.i = 0.5
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group 
hu.mackey.thomas.linear.power(n=180, t=t, 
  sig2.s=24, sig2.e=10, cor.s.i=0.5, p=p, power=0.80)
#> 
#>      Random Coefficients Regression Model. Hu, Mackey & Thomas (2021) 
#> 
#>               N = 360
#>               n = 180, 180
#>           delta = 1.49879
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>          sig2.i = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>         cor.s.i = 0.5
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group 
hu.mackey.thomas.linear.power(n=180, delta=1.5, t=t, 
  sig2.s=24, sig2.e=10, cor.s.i=0.5, p=p)
#> 
#>      Random Coefficients Regression Model. Hu, Mackey & Thomas (2021) 
#> 
#>               N = 360
#>               n = 180, 180
#>           delta = 1.5
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>          sig2.i = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>         cor.s.i = 0.5
#>       sig.level = 0.05
#>           power = 0.8006328
#>     alternative = two.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group 

hu.mackey.thomas.linear.power(delta=1.5, t=t, lambda=2, 
  sig2.s=24, sig2.e=10, cor.s.i=0.5, p=p, power=0.80)
#> 
#>      Random Coefficients Regression Model. Hu, Mackey & Thomas (2021) 
#> 
#>               N = 404.3467
#>               n = 269.5645, 134.7822
#>           delta = 1.5
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>          sig2.i = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>         cor.s.i = 0.5
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group 
hu.mackey.thomas.linear.power(n=270, t=t, lambda=2, 
  sig2.s=24, sig2.e=10, cor.s.i=0.5, p=p, power=0.80)
#> 
#>      Random Coefficients Regression Model. Hu, Mackey & Thomas (2021) 
#> 
#>               N = 405
#>               n = 270, 135
#>           delta = 1.49879
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>          sig2.i = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>         cor.s.i = 0.5
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group 
hu.mackey.thomas.linear.power(n=270, delta=1.5, t=t, lambda=2, 
  sig2.s=24, sig2.e=10, p=p, cor.s.i=0.5)
#> 
#>      Random Coefficients Regression Model. Hu, Mackey & Thomas (2021) 
#> 
#>               N = 405
#>               n = 270, 135
#>           delta = 1.5
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>          sig2.i = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>         cor.s.i = 0.5
#>       sig.level = 0.05
#>           power = 0.8006328
#>     alternative = two.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group 

hu.mackey.thomas.linear.power(delta=1.5, t=t, 
  sig2.s=24, sig2.e=10, cor.s.i=0.5, p=p, power=0.80, alternative='one.sided')
#> 
#>      Random Coefficients Regression Model. Hu, Mackey & Thomas (2021) 
#> 
#>               N = 283.1143
#>               n = 141.5572, 141.5572
#>           delta = 1.5
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>          sig2.i = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>         cor.s.i = 0.5
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = one.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group 
hu.mackey.thomas.linear.power(n=142, t=t, 
  sig2.s=24, sig2.e=10, cor.s.i=0.5, p=p, power=0.80, alternative='one.sided')
#> 
#>      Random Coefficients Regression Model. Hu, Mackey & Thomas (2021) 
#> 
#>               N = 284
#>               n = 142, 142
#>           delta = 1.497659
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>          sig2.i = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>         cor.s.i = 0.5
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = one.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group 
hu.mackey.thomas.linear.power(n=142, delta=1.5, t=t, 
  sig2.s=24, sig2.e=10, cor.s.i=0.5, p=p, sig.level=0.05, alternative='one.sided')
#> 
#>      Random Coefficients Regression Model. Hu, Mackey & Thomas (2021) 
#> 
#>               N = 284
#>               n = 142, 142
#>           delta = 1.5
#>               t = 0.00, 0.25, 0.50, 0.75, 1.00, 1.25, 1.50
#>               p = 0, 0, 0, 0, 0, 0, 1
#>          sig2.i = 0
#>          sig2.s = 24
#>          sig2.e = 10
#>         cor.s.i = 0.5
#>       sig.level = 0.05
#>           power = 0.8010862
#>     alternative = one.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group