Sample size calculations for difference in slopes between two groups.

This function performs the sample size calculation for difference in slopes between two groups. See Diggle, et al (2002) and package vignette for more details.

diggle.linear.power(
  n = NULL,
  delta = NULL,
  t = NULL,
  sigma2 = 1,
  R = NULL,
  sig.level = 0.05,
  power = NULL,
  alternative = c("two.sided", "one.sided"),
  tol = .Machine$double.eps^2
)

Arguments

n

sample size per group

delta

group difference in slopes

t

the observation times

sigma2

the residual variance

R

the working correlation matrix (or variance-covariance matrix if sigma2 is 1). If R is a scalar, an exchangeable working correlation matrix will be assumed.

sig.level

Type I error

power

power

alternative

one- or two-sided test

tol

numerical tolerance used in root finding.

Value

The number of subject required per arm to attain the specified power given sig.level and the other parameter estimates.

References

Diggle P.J., Heagerty P.J., Liang K., Zeger S.L. (2002) Analysis of longitudinal data. Second Edition. Oxford Statistical Science Series.

See also

Author

Michael C. Donohue, Steven D. Edland

Examples

if (FALSE) {
browseVignettes(package = "longpower")
}

# Reproduces the table on page 29 of Diggle et al
n <- 3
t <- c(0,2,5)
rho <- c(0.2, 0.5, 0.8)
sigma2 <- c(100, 200, 300)
tab <- outer(rho, sigma2, 
      Vectorize(function(rho, sigma2){
        ceiling(diggle.linear.power(
          delta=0.5,
          t=t,
          sigma2=sigma2,
          R=rho,
          alternative="one.sided",
          power = 0.80)$n[1])}))
colnames(tab) <- paste("sigma2 =", sigma2)
rownames(tab) <- paste("rho =", rho)
tab
#>           sigma2 = 100 sigma2 = 200 sigma2 = 300
#> rho = 0.2          313          625          938
#> rho = 0.5          196          391          586
#> rho = 0.8           79          157          235

# An Alzheimer's Disease example using ADAS-cog pilot estimates
# var of random intercept
sig2.i <- 55
# var of random slope
sig2.s <- 24
# residual var
sig2.e <- 10
# covariance of slope and intercep
cov.s.i <- 0.8*sqrt(sig2.i)*sqrt(sig2.s)

cov.t <- function(t1, t2, sig2.i, sig2.s, cov.s.i){
        sig2.i + t1*t2*sig2.s + (t1+t2)*cov.s.i 
}

t <- seq(0,1.5,0.25)
n <- length(t)
R <- outer(t, t, function(x,y){cov.t(x,y, sig2.i, sig2.s, cov.s.i)})
R <- R + diag(sig2.e, n, n)

diggle.linear.power(d=1.5, t=t, R=R, sig.level=0.05, power=0.80)
#> 
#>      Longitudinal linear model slope power calculation (Diggle et al 2002, page 29) 
#> 
#>               N = 414.6202
#>               n = 207.3101, 207.3101
#>           delta = 1.5
#>          sigma2 = 1
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 
#>  NOTE: N is *total* sample size and n is sample size in *each* group 
#> 
#>  R:
#>          [,1]      [,2]      [,3]      [,4]      [,5]     [,6]      [,7]
#> [1,] 65.00000  62.26636  69.53272  76.79908  84.06544  91.3318  98.59817
#> [2,] 62.26636  81.03272  79.79908  88.56544  97.33180 106.0982 114.86453
#> [3,] 69.53272  79.79908 100.06544 100.33180 110.59817 120.8645 131.13089
#> [4,] 76.79908  88.56544 100.33180 122.09817 123.86453 135.6309 147.39725
#> [5,] 84.06544  97.33180 110.59817 123.86453 147.13089 150.3972 163.66361
#> [6,] 91.33180 106.09817 120.86453 135.63089 150.39725 175.1636 179.92997
#> [7,] 98.59817 114.86453 131.13089 147.39725 163.66361 179.9300 206.19633
#>