longpower: Sample Size Calculations for Longitudinal Data

Compute power and sample size for linear models of longitudinal data. Supported models include mixed-effects models and models fit by generalized least squares and generalized estimating equations. The package is described in Iddi and Donohue (2022) doi:10.32614/RJ-2022-022 . Relevant formulas are derived by Liu and Liang (1997) doi:10.2307/2533554 , Diggle et al (2002) <ISBN:9780199676750>, and Lu, Luo, and Chen (2008) doi:10.2202/1557-4679.1098 .

See also

Author

Maintainer: Michael C. Donohue mdonohue@usc.edu

Other contributors:

Steve D. Edland sedland@health.ucsd.edu [contributor]

Nan Hu hu.nan@gene.com [contributor]