Linear mixed model sample size calculations.

This function performs the sample size calculation for a mixed model of repeated measures with general correlation structure. See Lu, Luo, & Chen (2008) for parameter definitions and other details. This function executes Formula (3) on page 4.

power.mmrm(
  N = NULL,
  Ra = NULL,
  ra = NULL,
  sigmaa = NULL,
  Rb = NULL,
  rb = NULL,
  sigmab = NULL,
  lambda = 1,
  delta = NULL,
  sig.level = 0.05,
  power = NULL,
  alternative = c("two.sided", "one.sided"),
  tol = .Machine$double.eps^2
)

Arguments

N

total sample size

Ra

correlation matrix for group a

ra

retention in group a

sigmaa

standard deviation of observation of interest in group a

Rb

correlation matrix for group a

rb

retention in group b

sigmab

standard deviation of observation of interest in group b. If NULL, sigmab is assumed same as sigmaa. If not NULL, sigmaa and sigmab are averaged.

lambda

allocation ratio

delta

effect size

sig.level

type one error

power

power

alternative

one- or two-sided test

tol

numerical tolerance used in root finding.

Value

The number of subject required per arm to attain the specified power given sig.level and the other parameter estimates.

Details

See Lu, Luo, & Chen (2008).

References

Lu, K., Luo, X., Chen, P.-Y. (2008) Sample size estimation for repeated measures analysis in randomized clinical trials with missing data. International Journal of Biostatistics, 4, (1)

See also

Author

Michael C. Donohue

Examples

# reproduce Table 1 from Lu, Luo, & Chen (2008)
phi1 <- c(rep(1, 6), 2, 2)
phi2 <- c(1, 1, rep(2, 6))
lambda <- c(1, 2, sqrt(1/2), 1/2, 1, 2, 1, 2)
ztest <- ttest1 <- c()
for(i in 1:8){
  Na <- (phi1[i] + lambda[i] * phi2[i])*(qnorm(0.05/2) + qnorm(1-0.90))^2*(0.5^-2)
  Nb <- Na/lambda[i]
  ztest <- c(ztest, Na + Nb)
  v <- Na + Nb - 2
  Na <- (phi1[i] + lambda[i] * phi2[i])*(qt(0.05/2, df = v) + qt(1-0.90, df = v))^2*(0.5^-2)
  Nb <- Na/lambda[i]
  ttest1 <- c(ttest1, Na + Nb)
}
data.frame(phi1, phi2, lambda, ztest, ttest1)
#>   phi1 phi2    lambda    ztest   ttest1
#> 1    1    1 1.0000000 168.1188 170.1476
#> 2    1    1 2.0000000 189.1336 191.1576
#> 3    1    2 0.7071068 244.9670 246.9822
#> 4    1    2 0.5000000 252.1782 254.1925
#> 5    1    2 1.0000000 252.1782 254.1925
#> 6    1    2 2.0000000 315.2227 317.2313
#> 7    2    2 1.0000000 336.2375 338.2447
#> 8    2    2 2.0000000 378.2672 380.2721

Ra <- matrix(0.25, nrow = 4, ncol = 4)
diag(Ra) <- 1
ra <- c(1, 0.90, 0.80, 0.70)
sigmaa <- 1
power.mmrm(Ra = Ra, ra = ra, sigmaa = sigmaa, delta = 0.5, power = 0.80)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 86.99175
#>              n2 = 86.99175
#>      retention1 = 1.0, 0.9, 0.8, 0.7
#>      retention2 = 1.0, 0.9, 0.8, 0.7
#>           delta = 0.5
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 
power.mmrm(N = 174, Ra = Ra, ra = ra, sigmaa = sigmaa, delta = 0.5)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 87
#>              n2 = 87
#>      retention1 = 1.0, 0.9, 0.8, 0.7
#>      retention2 = 1.0, 0.9, 0.8, 0.7
#>           delta = 0.5
#>       sig.level = 0.05
#>           power = 0.8000372
#>     alternative = two.sided
#> 
power.mmrm(N = 174, Ra = Ra, ra = ra, sigmaa = sigmaa, power = 0.80)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 87
#>              n2 = 87
#>      retention1 = 1.0, 0.9, 0.8, 0.7
#>      retention2 = 1.0, 0.9, 0.8, 0.7
#>           delta = 0.4999763
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 

power.mmrm(Ra = Ra, ra = ra, sigmaa = sigmaa, delta = 0.5, power = 0.80, lambda = 2)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 130.4876
#>              n2 = 65.24381
#>      retention1 = 1.0, 0.9, 0.8, 0.7
#>      retention2 = 1.0, 0.9, 0.8, 0.7
#>           delta = 0.5
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 
power.mmrm(N = 174, Ra = Ra, ra = ra, sigmaa = sigmaa, delta = 0.5, lambda = 2)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 116
#>              n2 = 58
#>      retention1 = 1.0, 0.9, 0.8, 0.7
#>      retention2 = 1.0, 0.9, 0.8, 0.7
#>           delta = 0.5
#>       sig.level = 0.05
#>           power = 0.7522291
#>     alternative = two.sided
#> 
power.mmrm(N = 174, Ra = Ra, ra = ra, sigmaa = sigmaa, power = 0.80, lambda = 2)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 116
#>              n2 = 58
#>      retention1 = 1.0, 0.9, 0.8, 0.7
#>      retention2 = 1.0, 0.9, 0.8, 0.7
#>           delta = 0.5303049
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 

# Extracting paramaters from gls objects with general correlation

# Create time index:
Orthodont$t.index <- as.numeric(factor(Orthodont$age, levels = c(8, 10, 12, 14)))
with(Orthodont, table(t.index, age))
#>        age
#> t.index  8 10 12 14
#>       1 27  0  0  0
#>       2  0 27  0  0
#>       3  0  0 27  0
#>       4  0  0  0 27

fmOrth.corSym <- gls( distance ~ Sex * I(age - 11), 
  Orthodont,
  correlation = corSymm(form = ~ t.index | Subject),
  weights = varIdent(form = ~ 1 | age) )
summary(fmOrth.corSym)$tTable
#>                            Value  Std.Error   t-value      p-value
#> (Intercept)           24.9371232 0.47286660 52.736064 7.204873e-77
#> SexFemale             -2.2717427 0.74083959 -3.066443 2.760864e-03
#> I(age - 11)            0.8268037 0.08221771 10.056272 5.037413e-17
#> SexFemale:I(age - 11) -0.3504390 0.12881040 -2.720580 7.640131e-03

C <- corMatrix(fmOrth.corSym$modelStruct$corStruct)[[1]]
sigmaa <- fmOrth.corSym$sigma * 
          coef(fmOrth.corSym$modelStruct$varStruct, unconstrained = FALSE)['14']
ra <- seq(1,0.80,length=nrow(C))
power.mmrm(N=100, Ra = C, ra = ra, sigmaa = sigmaa, power = 0.80)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 50
#>              n2 = 50
#>      retention1 = 1.0000000, 0.9333333, 0.8666667, 0.8000000
#>      retention2 = 1.0000000, 0.9333333, 0.8666667, 0.8000000
#>           delta = 1.324154
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 

# Extracting paramaters from gls objects with compound symmetric correlation

fmOrth.corCompSymm <- gls( distance ~ Sex * I(age - 11), 
  Orthodont,
  correlation = corCompSymm(form = ~ t.index | Subject),
  weights = varIdent(form = ~ 1 | age) )
summary(fmOrth.corCompSymm)$tTable
#>                            Value Std.Error   t-value      p-value
#> (Intercept)           24.8683158 0.4733953 52.531814 1.063204e-76
#> SexFemale             -2.1973966 0.7416680 -2.962777 3.779020e-03
#> I(age - 11)            0.7943130 0.0770109 10.314293 1.335155e-17
#> SexFemale:I(age - 11) -0.3155598 0.1206529 -2.615435 1.023614e-02

C <- corMatrix(fmOrth.corCompSymm$modelStruct$corStruct)[[1]]
sigmaa <- fmOrth.corCompSymm$sigma *
          coef(fmOrth.corCompSymm$modelStruct$varStruct, unconstrained = FALSE)['14']
ra <- seq(1,0.80,length=nrow(C))
power.mmrm(N=100, Ra = C, ra = ra, sigmaa = sigmaa, power = 0.80)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 50
#>              n2 = 50
#>      retention1 = 1.0000000, 0.9333333, 0.8666667, 0.8000000
#>      retention2 = 1.0000000, 0.9333333, 0.8666667, 0.8000000
#>           delta = 1.307974
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 

# Extracting paramaters from gls objects with AR1 correlation

fmOrth.corAR1 <- gls( distance ~ Sex * I(age - 11), 
  Orthodont,
  correlation = corAR1(form = ~ t.index | Subject),
  weights = varIdent(form = ~ 1 | age) )
summary(fmOrth.corAR1)$tTable
#>                            Value Std.Error   t-value      p-value
#> (Intercept)           24.9908804 0.4327748 57.745689 7.817009e-81
#> SexFemale             -2.3288881 0.6780278 -3.434797 8.530068e-04
#> I(age - 11)            0.7835499 0.1152669  6.797701 6.873260e-10
#> SexFemale:I(age - 11) -0.3008508 0.1805885 -1.665947 9.873342e-02

C <- corMatrix(fmOrth.corAR1$modelStruct$corStruct)[[1]]
sigmaa <- fmOrth.corAR1$sigma *
          coef(fmOrth.corAR1$modelStruct$varStruct, unconstrained = FALSE)['14']
ra <- seq(1,0.80,length=nrow(C))
power.mmrm(N=100, Ra = C, ra = ra, sigmaa = sigmaa, power = 0.80)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 50
#>              n2 = 50
#>      retention1 = 1.0000000, 0.9333333, 0.8666667, 0.8000000
#>      retention2 = 1.0000000, 0.9333333, 0.8666667, 0.8000000
#>           delta = 1.303892
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#> 
power.mmrm.ar1(N=100, rho = C[1,2], ra = ra, sigmaa = sigmaa, power = 0.80)
#> 
#>      Power for Mixed Model of Repeated Measures (Lu, Luo, & Chen, 2008) 
#> 
#>              n1 = 50
#>              n2 = 50
#>             rho = 0.6332579
#>      retention1 = 1.0000000, 0.9333333, 0.8666667, 0.8000000
#>      retention2 = 1.0000000, 0.9333333, 0.8666667, 0.8000000
#>            phi1 = 1.193581
#>            phi2 = 1.193581
#>           delta = 1.303892
#>           times = 1, 2, 3, 4
#>       sig.level = 0.05
#>           power = 0.8
#>     alternative = two.sided
#>