Print method for longitudinal data power calculation object

Print object of class "power.longtest" in nice layout.

# S3 method for class 'power.longtest'
print(x, ...)

Arguments

x

Object of class "power.longtest".

...

further arguments to be passed to or from methods.

Value

none

Details

A power.longtest object is just a named list of numbers and character strings, supplemented with method and note elements. The method is displayed as a title, the note as a footnote, and the remaining elements are given in an aligned ‘name = value’ format.

See also